source/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
psycopg2-binary==2.9.5
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.2.3